Create component for binding backfill section
