+++ src/components/editor/Bindings/Backfill/types.ts
@@ -1,14 +1,19 @@
 import { ReactNode } from 'react';
 
-export interface BackfillProps {
+export interface BackfillButtonProps {
     description: ReactNode;
     bindingIndex?: number;
+}
+
+export interface BackfillCountProps {
+    disabled?: boolean;
 }
 
 export interface BackfillDataflowOptionProps {
     disabled?: boolean;
 }
 
-export interface BackfillCountProps {
-    disabled?: boolean;
+export interface BackfillProps {
+    bindingIndex: number;
+    collectionEnabled: boolean;
 }
